% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateCitrusJuiceZest.R
\name{calculateCitrusJuiceZest}
\alias{calculateCitrusJuiceZest}
\title{calculateCitrusJuiceZest}
\usage{
calculateCitrusJuiceZest(df)
}
\arguments{
\item{df}{A dataframe with standardised ingredient and unit names.}
}
\value{
The dataframe with the amounts of citrus juice/zest from whole foods in actual amounts of juice/zest
}
\description{
Calculate the amounts of citrus juice and zest when recipe says to use t.ex "juice of one lemon".
}
\details{
Calculate the amounts of citrus juice and zest when recipe says to use t.ex "juice of one lemon".
}
