% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMeanAmounts.R
\name{calculateMeanAmounts}
\alias{calculateMeanAmounts}
\title{calculateMeanAmounts}
\usage{
calculateMeanAmounts(df)
}
\arguments{
\item{df}{A dataframe with an Ingredients column.}
}
\value{
The dataframe with the mean of these type of amounts.
}
\description{
Calculate the mean amount of ingredients when recipe says 3-4 and similar.
}
\details{
Calculate the mean amount of ingredients when recipe says 3-4 and similar.
}
