% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateNutrientContentOfFoodlist.R
\name{calculateNutrientContentOfFoodlist}
\alias{calculateNutrientContentOfFoodlist}
\title{calculateNutrientContentOfFoodlist}
\usage{
calculateNutrientContentOfFoodlist(
  df,
  calculate_nutrients = "total",
  identifier = "recipe_name"
)
}
\arguments{
\item{df}{The output from findFoodInDatabase "nutrients". To calculate nutrient content per 100 g amounts must be in hektograms.}

\item{calculate_nutrients}{Calculate nutrient content for the "total", "per 100 g" or "per portion". The default is for the "total"
of the list of foods. To calculate per 100 g, amounts of ingredients must be in hektograms. If "per portion" df must have a
"number_of_portions" column.}

\item{identifier}{A the column name for a column with an identifier for where the ingredients comes from. Default is "recipe_name".}
}
\value{
The nutrient content of each individual variable in the "identifier" column, either by total weight of all ingredients,
per 100 grams or per portion.
}
\description{
Calculate nutrient content of a list of foods with their amounts in hektograms.
}
\details{
Calculate nutrient content of a list of foods with their amounts in hektograms.
}
