% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateNutriscore.R
\name{calculateNutriscore}
\alias{calculateNutriscore}
\title{calculateNutriscore}
\usage{
calculateNutriscore(df, raw_scores = FALSE)
}
\arguments{
\item{df}{Dataframe with nutrient content per 100g and percentage of weight of fruit/vegetables and healthy oils formatted
similarly to the output of calculateNutrientContentOfFoodlist calculated "per 100 g". Fruit/vegetables and healthy oils
should be present in the "nutrient" column with the variable name "nutriscore_amounts_pct" and the value in percent should
be in the "nutrient_content_per_hektogram" column.}
}
\value{
The dataframe with the recipes raw nutriscore and equivalent nutriscore letter.
}
\description{
Calculate the nutriscore of a recipe.
}
\details{
Calculate the mean amount of ingredients when recipe says 3-4 and similar.
}
