% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFoodDatabaseMapping.R
\name{checkFoodDatabaseMapping}
\alias{checkFoodDatabaseMapping}
\title{checkFoodDatabaseMapping}
\usage{
checkFoodDatabaseMapping(df)
}
\arguments{
\item{df}{An output dataframe from findFoodInDatabase.}
}
\value{
A three column dataframe with the foods and their database reference with name and individual ID.
}
\description{
Check what reference foods from the database the foods have been mapped to.
}
