% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFoodStandardisation.R
\name{checkFoodStandardisation}
\alias{checkFoodStandardisation}
\title{checkFoodStandardisation}
\usage{
checkFoodStandardisation(df)
}
\arguments{
\item{df}{Output from standardiseFoodList function.}
}
\value{
A tibble with the unique values from the Original_ingredients and Ingredients column.
}
\description{
Look over the standardised names and amounts in a food list.
}
\details{
Look over the standardised names and amounts in a food list.
}
