% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{composite_ingredients_oda}
\alias{composite_ingredients_oda}
\title{39 composite ingredients found in the recipes from Oda.}
\format{
A tibble with 326 rows and two variables: recipe_name and Ingredients.
}
\usage{
composite_ingredients_oda
}
\description{
39 composite ingredients found in the recipes from Oda.
}
\keyword{datasets}
