% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCitrusJuiceZest.R
\name{convertCitrusJuiceZest}
\alias{convertCitrusJuiceZest}
\title{convertCitrusJuiceZest}
\usage{
convertCitrusJuiceZest(
  df,
  whole_fruit_to_volume_units = TRUE,
  weight_to_whole_fruits = FALSE
)
}
\arguments{
\item{df}{A dataframe with standardised ingredient and unit names.}

\item{whole_fruit_to_volume_units}{Convert whole fruits to volume units, f.ex. if a food list includes "juice of half a lemon". Default TRUE.}

\item{weight_to_whole_fruits}{Convert weight of juice and zest to whole fruits.}
}
\value{
The dataframe with the amounts of citrus juice/zest from whole foods in actual amounts of juice/zest
}
\description{
Convert volume amounts of citrus juice and zest when recipe says to use t.ex "juice of one lemon" from whole fruits to volume. Or from kilograms to kilogram of whole fruits.
}
\details{
Calculate the amounts of citrus juice and zest when recipe says to use t.ex "juice of one lemon".
}
