% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFreshHerbsToDried.R
\name{convertFreshHerbsToDried}
\alias{convertFreshHerbsToDried}
\title{convertFreshHerbsToDried}
\usage{
convertFreshHerbsToDried(df)
}
\arguments{
\item{df}{A dataframe with standardised ingredient names.}
}
\value{
A dataframe with the amounts of herbs converted to dried or fresh weight, and ingredient name changed from "fresh herbs" to "dried" or "dried" to "fresh herbs".
}
\description{
Convert the amounts of herbs from fresh to dried / dried to fresh if database only have the dried or fresh herbs.
}
\details{
Convert the amounts of herbs from fresh to dried if database only have the dried herbs.
}
