% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookedToRaw.R
\name{cookedToRaw}
\alias{cookedToRaw}
\title{cookedToRaw}
\usage{
cookedToRaw(df)
}
\arguments{
\item{df}{A dataframe standardised ingredient and unit names.}
}
\value{
The dataframe wwhere cooked ingredients have been trned into their raw equivalents.
}
\description{
Turn cooked ingredient amounts into their equivalents in raw form, using conversion factors from the Norwegian Directorate of Health.
}
\details{
Turn cooked ingredient amounts into their equivalents in raw form.
}
