% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixFoodMappingError.R
\name{fixFoodMappingError}
\alias{fixFoodMappingError}
\title{fixFoodMappingError}
\usage{
fixFoodMappingError(database, first_w, second_w = "\\\\")
}
\arguments{
\item{database}{Reference database to look up.}

\item{first_w}{The correct first word in reference database.}

\item{second_w}{The correct second word in reference database.}
}
\value{
The updated ID number.
}
\description{
Set a new ID number for an ingredient if findFoodInDatabase made a mistake.
}
\details{
Fix the ID number of an ingredient if findFoodInDatabase made a mistake.
}
