% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{matvaretabellen2020}
\alias{matvaretabellen2020}
\title{Nutrient database mainly from the 2020 edition of the Norwegian food composition datatable Matvaretabellen.}
\format{
A dataframe with 2009 rows and 45 variables.
\describe{
  \item{database_ID}{The unique identifier used for food items in Matvaretabellen, FoodData Central or generated by user}
  \item{from}{Where is the nutrient information from: Matvaretabellen, foodDataCentral, or a composite ingredient made by user}
}
}
\source{
\url{https://www.matportalen.no/verktoy/matvaretabellen/}

\url{https://fdc.nal.usda.gov/download-datasets.html}
}
\usage{
matvaretabellen2020
}
\description{
A dataset containing the nutrient content of more than 1000 foods from Matvaretabellen 2020 edition,
with Norwegian and English names. Also includes some foods not found in Matvaretabellen from the
American food composition datatable from FoodData Central April2021 and for computed composite ingredients.
}
\keyword{datasets}
