% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{matvaretabellen2020_query}
\alias{matvaretabellen2020_query}
\title{Query words used to search through and map recipe ingredients to matvaretabellen2020.
One or two words to map ingredients to matvaretabellen2020 dataframe.}
\format{
A dataframe with 611 rows and three variables.
\describe{
\item{first_word}{Generic ingredient name to identify ingredients}
\item{second_word}{A specifier to go with the generic word, f.ex. type of bean or cut of meat}
\item{database_ID}{The unique identifier used for food items in Matvaretabellen, FoodData Central or generated by user}
}
}
\usage{
matvaretabellen2020_query
}
\description{
Query words used to search through and map recipe ingredients to matvaretabellen2020.
One or two words to map ingredients to matvaretabellen2020 dataframe.
}
\keyword{datasets}
