% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{matvaretabellen2022_foodgroups}
\alias{matvaretabellen2022_foodgroups}
\title{Foodgroup categories from Matvaretabellen.}
\format{
A tibble with 2213 rows and two columns. The database_ID from Matvaretabellen
and the foodgroup this food belongs to in Matvaretabellen.
}
\usage{
matvaretabellen2022_foodgroups
}
\description{
Foodgroup categories from Matvaretabellen.
}
\keyword{datasets}
