% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{matvaretabellen_units}
\alias{matvaretabellen_units}
\title{The units for the nutrients from matvaretabellen}
\format{
A tibble containting a column for nutrients and a corresponding column with their units.
}
\usage{
matvaretabellen_units
}
\description{
The units for the nutrients from matvaretabellen
}
\keyword{datasets}
