% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutRients-package.R
\docType{package}
\name{nutRients-package}
\alias{nutRients}
\alias{nutRients-package}
\title{nutRients: Map foods to food composition and environmental impact databases}
\description{
Standardise food names in recipes or food diaries and map them to the Norwegian food composition datatable and the environmental impact database SHARP Indicator Database to calculate nutrient content and environmental impact.
}
\keyword{internal}
