% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcs}
\alias{pcs}
\title{A character vector of ingredients that when lacking a unit descriptor in recipe is
counted as a piece, ie "3 kiwis" will be interpreted as "3 pieces of kiwi".}
\format{
A character vector with 100 ingredient names.
}
\usage{
pcs
}
\description{
A character vector of ingredients that when lacking a unit descriptor in recipe is
counted as a piece, ie "3 kiwis" will be interpreted as "3 pieces of kiwi".
}
\keyword{datasets}
