% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recipes_oda}
\alias{recipes_oda}
\title{100 recipes scraped from the Norwegian online supermarket Oda.}
\format{
A list with two items
\itemize{
\item Recipes. A tibble with 935 rows and three columns, containing 100 recipes in total.
\itemize{
\item recipe_name. Individual recipe names.
\item Ingredients. Column with ingredient names and their amounts.
\item ingredient_is_part_of Metadata from Oda showing which part of the recipe the ingredient is from.
}
\item meta. A tibble with 654 rows and five columns, containing metadata for the recipes.
\itemize{
\item recipe_name. Individual recipe names.
\item number_of_portions. Number of portions in the recipe.
\item likes. Number of likes the recipe has received.
\item categories. Categories of the recipe from Oda.
}
}
}
\source{
\url{https://oda.com/no/recipes}
}
\usage{
recipes_oda
}
\description{
A list of two dataframes, one containing recipes and one containing
recipe metadata such as number of portions and how the recipe
was categorised by Oda.
}
\keyword{datasets}
