% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SHARP2018}
\alias{SHARP2018}
\title{The environmental impact database SHARP Identificator Database}
\format{
A dataframe with 600 rows and five variables.
\describe{
\item{FoodEx2}{A FoodEx2 unique identifier for the food}
\item{L1}{The food group the food contains}
\item{GHGE of 1 kg food as consumed_kgCO2eq}{Greenhouse gas emissions of 1 kg food as consumed, in kilo CO2 equivalents}
\item{Land use of 1 kg food as consumed_m2/yr}{Land used to produce 1 kg of food as consumed, in m2/year}
\item{database_ID}{A unique ID for the food used for internal functions}
}
}
\source{
\url{https://easy.dans.knaw.nl/ui/datasets/id/easy-dataset:132158}
}
\usage{
SHARP2018
}
\description{
A dataset containing 600 unique foods and their environmental impact in
greenhouse gas emission and land use.
}
\keyword{datasets}
