% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SHARP2018_foodgroups}
\alias{SHARP2018_foodgroups}
\title{Foodgroup categories from SHARP ID.}
\format{
A tibble with 600 rows and two columns. The database_ID from SHARP ID
and the foodgroup this food belongs to in SHARP ID.
}
\usage{
SHARP2018_foodgroups
}
\description{
Foodgroup categories from SHARP ID.
}
\keyword{datasets}
