% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SHARP2018_query}
\alias{SHARP2018_query}
\title{Query words used to search through and map recipe ingredients to SHARP2018.}
\format{
A dataframe with 578 rows and three variables.
\describe{
\item{first_word}{Generic ingredient name to identify ingredients}
\item{second_word}{A specifier to go with the generic word, f.ex. type of bean or cut of meat}
\item{database_ID}{A unique ID used to map query words to SHARP2018}
}
}
\usage{
SHARP2018_query
}
\description{
One or two words to map ingredients to the SHARP2018 dataframe.
}
\keyword{datasets}
