% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitBrothtoWaterBrotchcubes.R
\name{splitBrothtoWaterBrotchcubes}
\alias{splitBrothtoWaterBrotchcubes}
\title{splitBrothtoWaterBrotchcubes}
\usage{
splitBrothtoWaterBrotchcubes(df)
}
\arguments{
\item{df}{A dataframe with a standardised Ingredients column.}
}
\value{
A dataframe with the broth ingredients split into the equivalent amounts of water and broth cubes needed to make the broth.
}
\description{
Split broth into water and broth cubes, as these ingredients are found in the databases not broth.
}
\details{
Split broth into water and broth cubes, as these ingredients are found in the databases not broth.
}
