% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardiseAmounts.R
\name{standardiseAmounts}
\alias{standardiseAmounts}
\title{standardiseAmounts}
\usage{
standardiseAmounts(df)
}
\arguments{
\item{df}{A dataframe with an Ingredients column.}
}
\value{
The dataframe with amounts standardised.
}
\description{
Standardise different ways of spelling amounts and units used of an ingredient.
}
\details{
Standardise amounts of an ingredient.
}
