% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardiseBeverages.R
\name{standardiseBeverages}
\alias{standardiseBeverages}
\title{standardiseBeverages}
\usage{
standardiseBeverages(df)
}
\arguments{
\item{df}{A dataframe with an Ingredients column, listing each ingredient of the recipe in individual rows, and an Ingredients_standardised column.}
}
\value{
The dataframe with a column with beverages with standardised names.
}
\description{
Support function for "standardiseFoodNames", standardises names of different beverages
}
\details{
Standardising ingredients names in a recipe, here different beverages
}
