% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardiseCondiments.R
\name{standardiseCondiments}
\alias{standardiseCondiments}
\title{standardiseCondiments}
\usage{
standardiseCondiments(df)
}
\arguments{
\item{df}{A dataframe with an Ingredients column, listing each ingredient of the recipe in individual rows, and an Ingredients_standardised column.}
}
\value{
The dataframe with a column with various other foods with standardised names.
}
\description{
Standardise names of various condiments
}
\details{
Standardising ingredients names in a recipe, here various condiments of foods.
}
