% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardiseDairynSubs.R
\name{standardiseDairynSubs}
\alias{standardiseDairynSubs}
\title{standardiseDairynSubs}
\usage{
standardiseDairynSubs(df)
}
\arguments{
\item{df}{A dataframe with an Ingredients column, listing each ingredient of the recipe in individual rows, and an Ingredients_standardised column.}
}
\value{
The dataframe with a column with dairy and substitutes standardised.
}
\description{
Standardise names of dairy and dairy substitute.
}
\details{
Standardising ingredients names in a recipe, here dairy and substitutes.
}
