% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardiseDairynSubsB.R
\name{standardiseDairynSubsB}
\alias{standardiseDairynSubsB}
\title{standardiseDairynSubsB}
\usage{
standardiseDairynSubsB(df)
}
\arguments{
\item{df}{A dataframe with an Ingredients column, listing each ingredient of the recipe in individual rows, and an Ingredients_standardised column.}
}
\value{
The dataframe with a column with dairy and substitutes starting with the letter "B" with standardised names.
}
\description{
Support function for "standardiseDairynSubs", standardise names of dairy and dairy substitute starting with "B".
}
\details{
Standardising ingredients names in a recipe, here dairy and substitutes starting with the letter "B".
}
