% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardiseDairynSubsS.R
\name{standardiseDairynSubsS}
\alias{standardiseDairynSubsS}
\title{standardiseDairynSubsS}
\usage{
standardiseDairynSubsS(df)
}
\arguments{
\item{df}{A dataframe with an Ingredients column, listing each ingredient of the recipe in individual rows, and an Ingredients_standardised column.}
}
\value{
The dataframe with a column with dairy and substitutes starting with the letter "S" with standardised names.
}
\description{
Support function for "standardiseDairynSubs", standardise names of dairy and dairy substitute starting with "S".
}
\details{
Standardising ingredients names in a recipe, here dairy and substitutes starting with the letter "S".
}
