% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardiseFlournGraisnNutsnLegumesT.R
\name{standardiseFlournGraisnNutsnLegumesT}
\alias{standardiseFlournGraisnNutsnLegumesT}
\title{standardiseFlournGraisnNutsnLegumesT}
\usage{
standardiseFlournGraisnNutsnLegumesT(df)
}
\arguments{
\item{df}{A dataframe with an Ingredients column, listing each ingredient of the recipe in individual rows, and an Ingredients_standardised column.}
}
\value{
The dataframe with a column with flours, grains nuts and legumes starting with the letter "T" with standardised names.
}
\description{
Support function for "standardiseFlournGraisnNutsnLegumes", standardise names of flours, grains nuts and legumes starting with "T".
}
\details{
Standardising ingredients names in a recipe, here flours, grains nuts and legumes starting with the letter "T".
}
