% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardiseFlournGraisnNutsnLegumesW.R
\name{standardiseFlournGraisnNutsnLegumesW}
\alias{standardiseFlournGraisnNutsnLegumesW}
\title{standardiseFlournGraisnNutsnLegumesW}
\usage{
standardiseFlournGraisnNutsnLegumesW(df)
}
\arguments{
\item{df}{A dataframe with an Ingredients column, listing each ingredient of the recipe in individual rows, and an Ingredients_standardised column.}
}
\value{
The dataframe with a column with flours, grains nuts and legumes starting with the letter "W" with standardised names.
}
\description{
Support function for "standardiseFlournGraisnNutsnLegumes", standardise names of flours, grains nuts and legumes starting with "W".
}
\details{
Standardising ingredients names in a recipe, here flours, grains nuts and legumes starting with the letter "W".
}
