% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardiseFoodNames.R
\name{standardiseFoodNames}
\alias{standardiseFoodNames}
\title{standardiseFoodNames}
\usage{
standardiseFoodNames(df)
}
\arguments{
\item{df}{A dataframe with an Ingredients column, listing each food in an individual rows, an unit column noting which unit the food has, and an Ingredients_standardised column.}
}
\value{
The dataframe with a column with foods with standardised names.
}
\description{
Standardise names of all types of foods, running all the functions for standardising names of individual categories of food.
}
\details{
Standardising food names in a recipe.
}
