% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardiseFruitVeg.R
\name{standardiseFruitVeg}
\alias{standardiseFruitVeg}
\title{standardiseFruitVeg}
\usage{
standardiseFruitVeg(df)
}
\arguments{
\item{df}{A dataframe with an Ingredients column, listing each ingredient of the recipe in individual rows, and an Ingredients_standardised column.}
}
\value{
The dataframe with a new with fruit and vegetables with standardised names in the "Ingredients_standardised" column.
}
\description{
Standardise names of fruit and vegetables.
}
\details{
Standardising ingredients names in a recipe, here all fruit and vegetables.
}
