% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardiseFruitVegN.R
\name{standardiseFruitVegN}
\alias{standardiseFruitVegN}
\title{standardiseFruitVegN}
\usage{
standardiseFruitVegN(df)
}
\arguments{
\item{df}{A dataframe with an Ingredients column, listing each ingredient of the recipe in individual rows, and an Ingredients_standardised column.}
}
\value{
The dataframe with a new with fruit and vegetables starting with the letter "N" with standardised names.
}
\description{
Support function for "standardiseFruitVeg", standardise names of fruit and veg starting with "N".
}
\details{
Standardising ingredients names in a recipe, here fruit and vegetables starting with the letter "N".
}
