% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardiseHerbsnSpicesA.R
\name{standardiseHerbsnSpicesA}
\alias{standardiseHerbsnSpicesA}
\title{standardiseHerbsnSpicesA}
\usage{
standardiseHerbsnSpicesA(df)
}
\arguments{
\item{df}{A dataframe with an Ingredients column, listing each ingredient of the recipe in individual rows, and an Ingredients_standardised column.}
}
\value{
The dataframe with a column with herbs and spices starting with the letter "A" with standardised names.
}
\description{
Support function for "standardiseHerbsnSpices", standardise names of herbs and spices starting with "A".
}
\details{
Standardising ingredients names in a recipe, here herbs and spices starting with the letter "A".
}
