% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardiseMushrooms.R
\name{standardiseMushrooms}
\alias{standardiseMushrooms}
\title{standardiseMushrooms}
\usage{
standardiseMushrooms(df)
}
\arguments{
\item{df}{A dataframe with an Ingredients column, listing each ingredient of the recipe in individual rows, and an Ingredients_standardised column.}
}
\value{
The dataframe with a column with mushrooms with standardised names.
}
\description{
Standardise names of mushrooms.
}
\details{
Standardising ingredients names in a recipe, here mushrooms.
}
