% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardiseOils.R
\name{standardiseOils}
\alias{standardiseOils}
\title{standardiseOils}
\usage{
standardiseOils(df)
}
\arguments{
\item{df}{A dataframe with an Ingredients column, listing each ingredient of the recipe in individual rows, and an Ingredients_standardised column.}
}
\value{
The dataframe with a column with oils with standardised names.
}
\description{
Standardise names of oils
}
\details{
Standardising ingredients names in a recipe, here oils.
}
