% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardisePoultry.R
\name{standardisePoultry}
\alias{standardisePoultry}
\title{standardisePoultry}
\usage{
standardisePoultry(df)
}
\arguments{
\item{df}{A dataframe with an Ingredients column, listing each ingredient of the recipe in individual rows, and an Ingredients_standardised column.}
}
\value{
The dataframe with a column with poultry with standardised names.
}
\description{
Standardise names of poultry.
}
\details{
Standardising ingredients names in a recipe, here poultry.
}
