% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardiseRedMeatnSubs.R
\name{standardiseRedMeatnSubs}
\alias{standardiseRedMeatnSubs}
\title{standardiseRedMeatnSubs}
\usage{
standardiseRedMeatnSubs(df)
}
\arguments{
\item{df}{A dataframe with an Ingredients column, listing each ingredient of the recipe in individual rows, and an Ingredients_standardised column.}
}
\value{
The dataframe with a column with red meat with standardised names.
}
\description{
Standardise names of red meat and meat substitutions
}
\details{
Standardising ingredients names in a recipe, here red meat.
}
