% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardiserFoodNames.R
\name{standardiserFoodNames}
\alias{standardiserFoodNames}
\title{standardiserFoodNames
Standardise names of food items in Norwegian.}
\usage{
standardiserFoodNames(df)
}
\arguments{
\item{df}{Dataframe with an FoodItem column with food item names to standardise.}
}
\value{
The dataframe with an additional standardisedFoodItem column with standardised food item names.
}
\description{
standardiserFoodNames
Standardise names of food items in Norwegian.
}
