% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardiseSauces.R
\name{standardiseSauces}
\alias{standardiseSauces}
\title{standardiseSauces}
\usage{
standardiseSauces(df)
}
\arguments{
\item{df}{A dataframe with an Ingredients column, listing each ingredient of the recipe in individual rows, and an Ingredients_standardised column.}
}
\value{
The dataframe with a column with various other foods with standardised names.
}
\description{
Standardise names of various sauces
}
\details{
Standardising ingredients names in a recipe, here various sauces of foods.
}
