% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardiseSeafood.R
\name{standardiseSeafood}
\alias{standardiseSeafood}
\title{standardiseSeafood}
\usage{
standardiseSeafood(df)
}
\arguments{
\item{df}{A dataframe with an Ingredients column, listing each ingredient of the recipe in individual rows, and an Ingredients_standardised column.}
}
\value{
The dataframe with a column with seafood with standardised names.
}
\description{
Standardise names of seafood.
}
\details{
Standardising ingredients names in a recipe, here seafood.
}
