% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sustainableNutRients-package.R
\docType{package}
\name{sustainableNutRients-package}
\alias{sustainableNutRients}
\alias{sustainableNutRients-package}
\title{sustainableNutRients: Map foods to Matvaretabellen food composition and SHARP ID environmental impact databases}
\description{
Standardise food names in recipes or food diaries and map them to the Norwegian food composition datatable Matvaretabellen and the environmental impact database SHARP Indicator Database to calculate nutrient content and environmental impact.
}
\keyword{internal}
