% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{type_of_units}
\alias{type_of_units}
\title{A character vector of different unit names to extract from ingredient column.}
\format{
A character vector with 33 different unit names.
}
\usage{
type_of_units
}
\description{
A character vector of different unit names to extract from ingredient column.
}
\keyword{datasets}
