% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unit_weights}
\alias{unit_weights}
\title{Weight in grams of different units of foods.}
\format{
A data frame with 1685 rows and six variables.
\describe{
\item{Ingredients}{Ingredient name}
\item{unit_enhet}{The unit of the ingredient}
\item{grams_per_unit}{The weight of the unit of the ingredient in grams}
\item{reference}{The reference from where the weight per unit comes from}
\item{database_ID}{A unique identifier for the ingredient}
\item{language}{The language used for the ingredient and unit names}
}
}
\source{
\url{https://www.matportalen.no/verktoy/kostholdsplanleggeren/maal_vekt_og_porsjonsstorrelser_for_matvarer-1}
}
\usage{
unit_weights
}
\description{
A dataset containing the unit weights of multiple units
for more than 500 different items of food. The main source
for this dataset is the 2015 "Weights, measures and portion sizes for foods"
database from the Norwegian Directorate of Health, the University of Oslo
and the Norwegian Food Safety Authority.
}
\keyword{datasets}
