% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unit_weights_query}
\alias{unit_weights_query}
\title{Query words used to search through and map recipe ingredients to unit_weights database.}
\format{
A dataframe with 989 rows and four variables.
\describe{
  \item{database_ID}{A unique ID for the ingredient to link it to unit_weight database}
  \item{language}{The language used for the ingredient and unit names}
  \item{first_word}{Generic ingredient name to identify ingredients}
  \item{second_word}{A specifier to go with the generic word, f.ex. type of bean or cut of meat}
}
}
\usage{
unit_weights_query
}
\description{
One or two words to map ingredients to unit_weights dataframe
}
\keyword{datasets}
