% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardiseFoodList.R
\name{standardiseFoodList}
\alias{standardiseFoodList}
\title{standardiseFoodList}
\usage{
standardiseFoodList(df)
}
\arguments{
\item{df}{A dataframe with an Ingredients column.}
}
\value{
A dataframe with the amounts and standardised units of foods in their own column, a new Ingredients column with standardised
 names and an Original_ingredients column with the original ingredient names.
}
\description{
Standardises the names of foods and their units.
}
\details{
Standardise a list of foods into standardised food names and units.
}
